third_party/WebKit/LayoutTests/fast/dom/createDocumentType-ownerDocument-expected.txt

Tests that DOMImplementation.createDocumentType() properly sets the node's document to the associated document of the context object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS docType.ownerDocument is document
PASS newDocument = document.implementation.createDocument('', null, docType) did not throw exception.
PASS newDocument.doctype is docType
PASS newDocument.doctype.ownerDocument is newDocument
PASS successfullyParsed is true

TEST COMPLETE
